services/datalake_rest_api/requirements.txt
anyio==3.5.0
azure-core==1.22.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==3.1.1
fastapi==0.73.0
idna==3.3
install==1.3.5
isodate==0.6.1
msrest==0.6.21
numpy==1.21.5
oauthlib==3.2.0
pandas==1.3.5
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
urllib3==1.26.8
